Moving compress and decompress
