+++ lib/connection/connection.js
@@ -8,7 +8,7 @@ var inherits = require('util').inherits
   , f = require('util').format
   , debugOptions = require('./utils').debugOptions
   , parseHeader = require('../wireprotocol/shared').parseHeader
-  , compressorIDs = require('../wireprotocol/compression').compressorIDs
+  , decompress = require('../wireprotocol/compression').decompress
   , Response = require('./commands').Response
   , MongoError = require('../error')
   , Logger = require('./logger')
@@ -218,23 +218,6 @@ var closeHandler = function(self) {
         , MongoError.create(f("connection %s to %s:%s closed", self.id, self.host, self.port))
         , self);
     }
-  }
-}
-
-// Decompress a message using the given compressor
-var decompress = function(compressorID, compressedData, callback) {
-  if (compressorID < 0 || compressorID > compressorIDs.length) {
-    throw new Error('Server sent message compressed using an unsupported compressor. (Received compressor ID ' + compressorID + ')');
-  }
-  switch (compressorID) {
-    case compressorIDs.snappy:
-      Snappy.uncompress(compressedData, callback);
-      break;
-    case compressorIDs.zlib:
-      zlib.inflate(compressedData, callback);
-      break;
-    default:
-      callback(null, compressedData);
   }
 }
