+++ lib/connection/pool.js
@@ -13,6 +13,7 @@ var inherits = require('util').inherits,
   zlib = require('zlib'),
   MESSAGE_HEADER_SIZE = require('../wireprotocol/shared').MESSAGE_HEADER_SIZE,
   opcodes = require('../wireprotocol/shared').opcodes,
+  compress = require('../wireprotocol/compression').compress,
   compressorIDs = require('../wireprotocol/compression').compressorIDs,
   uncompressibleCommands = require('../wireprotocol/compression').uncompressibleCommands;
 
@@ -922,25 +923,6 @@ Pool.prototype.destroy = function(force) {
 
   // Initiate drain of operations
   checkStatus();
-}
-
-// Facilitate compressing a message using an agreed compressor
-var compress = function(self, dataToBeCompressed, callback) {
-  switch (self.options.agreedCompressor) {
-    case 'snappy':
-      Snappy.compress(dataToBeCompressed, callback);
-      break;
-    case 'zlib':
-      // Determine zlibCompressionLevel
-      var zlibOptions = {};
-      if (self.options.zlibCompressionLevel) {
-        zlibOptions.level = self.options.zlibCompressionLevel
-      }
-      zlib.deflate(dataToBeCompressed, zlibOptions, callback);
-      break;
-    default:
-      throw new Error('Attempt to compress message using unknown compressor \"' + self.options.agreedCompressor + '\".');
-  }
 }
 
 /**
